Move security context logic to k8s package
