+++ handlers/update.go
@@ -69,8 +69,8 @@ func updateDeploymentSpec(
 
 		deployment.Spec.Template.Spec.Containers[0].Env = buildEnvVars(&request)
 
-		configureReadOnlyRootFilesystem(request, deployment)
-		configureContainerUserID(deployment, nonRootFunctionuserID, factory.Config)
+		factory.ConfigureReadOnlyRootFilesystem(request, deployment)
+		factory.ConfigureContainerUserID(deployment)
 
 		deployment.Spec.Template.Spec.NodeSelector = createSelector(request.Constraints)
